# Fix potential deadlock with nested GetOrCreate calls