--- a/.github/workflows/test.yml
+++ b/.github/workflows/test.yml
@@ -18,9 +18,6 @@ jobs:
     - name: Install staticcheck
       run: go install honnef.co/go/tools/cmd/staticcheck@latest
       shell: bash
-    - name: Install golint
-      run: go install golang.org/x/lint/golint@latest
-      shell: bash
     - name: Update PATH
       run: echo "$(go env GOPATH)/bin" >> $GITHUB_PATH
       shell: bash
@@ -34,8 +31,6 @@ jobs:
       run: go vet ./...
     - name: Staticcheck
       run: staticcheck ./...
-    - name: Lint
-      run: golint ./...
     - name: Test
       run: go test -race ./... -coverpkg=./... -coverprofile=coverage.txt -covermode=atomic
     - name: Upload coverage