The new revived lock logic is also faster:

```
name               old time/op    new time/op    delta
NewGetOrCreate-10     301µs ± 0%     101µs ± 0%  -66.59%  (p=0.004 n=6+5)

name               old alloc/op   new alloc/op   delta
NewGetOrCreate-10     19.3B ± 3%     19.7B ± 3%     ~     (p=0.567 n=6+6)

name               old allocs/op  new allocs/op  delta
NewGetOrCreate-10      0.00           0.00          ~     (all equal)
```